Use aws-sdk as dev dependency
aws-sdk is already included in aws lambda so there
is no need to include it the build (dist) archive.

Also restructured to only copy package-lock.json
and package.json to build dir and do npm production
install from that directory (to avoid any clashes
with the lock file). Would like to use a production
version of the npm ci command, but that is currently
missing (https://github.com/npm/npm/issues/20125)